Refactoring Handler to separate command and query modules.
