--- src/server/vm/command.go
@@ -1,10 +1,0 @@
-package vm
-
-type Command interface {
-
-	Id() Identifier
-
-	TypeId() Identifier
-
-	AggregateTypeId() Identifier
-}
