+++ src/server/vm/command/player.go
@@ -1,4 +1,4 @@
-package handler
+package command
 
 type Player interface {
